Implement trait ModelAttribute::class.
