+++ src/Widget/Password.php
@@ -6,7 +6,9 @@ namespace Yiisoft\Form\Widget;
 
 use InvalidArgumentException;
 use Yiisoft\Form\Widget\Attribute\CommonAttribute;
+use Yiisoft\Form\Widget\Attribute\ModelAttribute;
 use Yiisoft\Html\Tag\Input;
+use Yiisoft\Widget\Widget;
 
 /**
  * The input element with a type attribute whose value is "password" represents a one-line plain-text edit control for
@@ -17,6 +19,7 @@ use Yiisoft\Html\Tag\Input;
 final class Password extends Widget
 {
     use CommonAttribute;
+    use ModelAttribute;
 
     /**
      * Specifies the form element the tag input element belongs to. The value of this attribute must be the id
