+++ src/Widget/RadioList.php
@@ -8,14 +8,18 @@ use Closure;
 use InvalidArgumentException;
 use Stringable;
 use Yiisoft\Arrays\ArrayHelper;
+use Yiisoft\Form\Widget\Attribute\ModelAttribute;
 use Yiisoft\Html\Widget\RadioList\RadioItem;
 use Yiisoft\Html\Widget\RadioList\RadioList as RadioListTag;
+use Yiisoft\Widget\Widget;
 
 /**
  * Generates a list of radio.
  */
 final class RadioList extends Widget
 {
+    use ModelAttribute;
+
     private array $containerAttributes = [];
     private ?string $containerTag = 'div';
     /** @psalm-var array<array-key, array<array-key, mixed>> */
