+++ src/Widget/ResetButton.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Yiisoft\Form\Widget;
 
+use Yiisoft\Form\Widget\Attribute\CommonAttribute;
+use Yiisoft\Form\Widget\Attribute\WithoutModelAttribute;
 use Yiisoft\Html\Tag\Input;
 use Yiisoft\Widget\Widget;
 
@@ -14,100 +16,8 @@ use Yiisoft\Widget\Widget;
  */
 final class ResetButton extends Widget
 {
-    private string $autoIdPrefix = 'reset-';
-    private array $attributes = [];
-    private string $id = '';
-    private string $name = '';
-    private string $value = '';
-    private static int $counter = 0;
-
-    /**
-     * The prefix to the automatically generated widget IDs.
-     *
-     * @param string $value
-     *
-     * @return static
-     */
-    public function autoIdPrefix(string $value): self
-    {
-        $new = clone $this;
-        $new->autoIdPrefix = $value;
-        return $new;
-    }
-
-    /**
-     * The HTML attributes. The following special options are recognized.
-     *
-     * @param array $value
-     *
-     * @return static
-     *
-     * See {@see \Yiisoft\Html\Html::renderTagAttributes()} for details on how attributes are being rendered.
-     */
-    public function attributes(array $value): self
-    {
-        $new = clone $this;
-        $new->attributes = $value;
-        return $new;
-    }
-
-    /**
-     * Set the Id of the widget.
-     *
-     * @param string $value
-     *
-     * @return static
-     *
-     * @link https://html.spec.whatwg.org/multipage/dom.html#the-id-attribute
-     */
-    public function id(string $id): self
-    {
-        $new = clone $this;
-        $new->id = $id;
-        return $new;
-    }
-
-    /**
-     * The name part of the name/value pair associated with this element for the purposes of form submission.
-     *
-     * @param string The name of the widget.
-     *
-     * @return static
-     *
-     * @link https://www.w3.org/TR/2012/WD-html-markup-20120329/input.reset.html#input.reset.attrs.name
-     */
-    public function name(string $value): self
-    {
-        $new = clone $this;
-        $new->name = $value;
-        return $new;
-    }
-
-    /**
-     * Specifies a value for the input element.
-     *
-     * @param string $value
-     *
-     * @return static
-     *
-     * @link https://www.w3.org/TR/2012/WD-html-markup-20120329/input.reset.html#input.reset.attrs.value
-     */
-    public function value(string $value): self
-    {
-        $new = clone $this;
-        $new->value = $value;
-        return $new;
-    }
-
-    /**
-     * Counter used to generate {@see id} for widgets.
-     *
-     * @param int $value
-     */
-    public static function counter(int $value): void
-    {
-        self::$counter = $value;
-    }
+    use CommonAttribute;
+    use WithoutModelAttribute;
 
     /**
      * @return string the generated input tag.
@@ -117,13 +27,14 @@ final class ResetButton extends Widget
         $new = clone $this;
         $input = Input::tag()->type('reset');
 
-        $id = $new->id !== '' ? $new->id : $new->autoIdPrefix . self::$counter++;
-        $name = $new->name !== '' ? $new->name : $id;
+        if ($new->autoIdPrefix === '') {
+            $new->autoIdPrefix = 'reset-';
+        }
 
         if ($new->value !== '') {
             $input = $input->value($new->value);
         }
 
-        return $input ->attributes($new->attributes)->id($id)->name($name)->render();
+        return $input ->attributes($new->attributes)->id($new->getId())->name($new->getName())->render();
     }
 }
